+++ packages/authhero/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Minor Changes
 
+- Only enforce ip check on magic link flow
 - Updated packages and added danish
 
 ## 0.158.0